Added the base of our game or rather skeleton or something like that
The main file is where we run the game I call it the launcher.

The entities file is where the object or characters of our game is 

The setting file is where we add the editable variables of our game. 

The utilities file is the functions that will help us with our code self 
made functions.

And the assests folder is where our assests is